+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # ChangeLog
 
+### Release_1.0.0_20200203_build_B
+
+#### 功能构建
+
+- ProfileService更名为ProfileMaintainService。
+
+#### Bug修复
+
+- (无)
+
+#### 功能移除
+
+- (无)
+
+---
+
 ### Release_1.0.0_20200203_build_A
 
 #### 功能构建